pyproject.toml
[project]
name = "social-content-creator"
version = "0.1.0"
description = "AIエンジニア「Aidotters」の情報発信効率化ツール"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-frontmatter>=1.1",
    "jinja2>=3.1",
    "markdown>=3.10.2",
    "python-dotenv>=1.0",
    "google-genai>=1.63.0",
    "authlib>=1.3",
]


[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "frontmatter"
ignore_missing_imports = true
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "authlib.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "respx>=0.22.0",
    "ruff>=0.8",
    "black>=24.0",
    "mypy>=1.14",
    "pytest-cov>=7.0.0",
]
